+++ src/renderer/data/utils.ts
@@ -273,7 +273,7 @@ export function extractMusicMetadata(audio: Audio, metadata: any, cachePath: str
       audio.trackNum = parseInt(metadata.common.track.no);
     }
     if (metadata.common.bpm) {
-      audio.bpm = parseInt(metadata.common.bpm);
+      audio.bpm = parseFloat(metadata.common.bpm);
     }
   }
   if (metadata.format && metadata.format.duration) {
